+++ src/registerCommands.ts
@@ -2,6 +2,8 @@ import * as vscode from 'vscode';
 import { Base64HelperUi } from './tools/base64Helper/base64HelperUi';
 import { JsonToCodeUi } from './tools/jsonToCode/jsonToCodeUi';
 import { CodeToJsonUi } from './tools/codeToJson/codeToJsonUi';
+import { XmlJsonUi } from './tools/xmlJson/xmlJsonUi';
+import { JwtDebuggerUi } from './tools/jwtDebugger/jwtDebuggerUi';
 
 export function registerCommands(context: vscode.ExtensionContext) {
 	context.subscriptions.push(
@@ -15,6 +17,14 @@ export function registerCommands(context: vscode.ExtensionContext) {
 		}),
 		vscode.commands.registerCommand('codemate.codeToJson', async () => {
 			const ui = new CodeToJsonUi();
+			await ui.show(context);
+		}),
+		vscode.commands.registerCommand('codemate.xmlJson', async () => {
+			const ui = new XmlJsonUi();
+			await ui.show(context);
+		}),
+		vscode.commands.registerCommand('codemate.jwtDebugger', async () => {
+			const ui = new JwtDebuggerUi();
 			await ui.show(context);
 		})
 	);